refactor(web): move plugin-specific connection-config assembly behind a plugin slot
Introduce PlatformContribution.connectionConfig (#1330) - the non-render
half of a platform's structured connection-config editing: a Zod schema
fragment composed into the edit-connection resolver at render time, a
read-side readConfigToForm hydrator, and a write-side applyToConfig
partial-patch assembler. Field names are declaration-merged into the new
PluginEditConnectionFields interface (mirrors the PluginApiNamespaces
precedent) so plugin sections stay statically typed.

Migrate KSeF as the first platform: every ksef-named module leaves the
shared features/connections feature (seller/payment assembly, NIP/NRB
normalizers, setup wizard schema+form+page, value sets) and lands in the
plugins/ksef slice; the KSeF Zod fields, superRefine checks, merge
clauses, and hydration readers move verbatim into
plugins/ksef/ksef-connection-config.ts, preserving the #1311
per-keystroke partial-patch semantics (relocated tests pin them).
Other platforms' inline fields migrate as follow-ups now that the seam
exists.

Closes #1330
