+++ apps/web/src/features/connections/components/edit-connection.schema.test.ts
@@ -1,7 +1,12 @@
 import { describe, expect, it } from 'vitest';
-import { editConnectionSchema, mergeStructuredIntoConfig } from './edit-connection.schema';
-import { buildKsefSellerConfig } from './ksef-setup.schema';
-import type { KsefSellerProfileInput } from './ksef-seller-config';
+import { z } from 'zod';
+import {
+  buildEditConnectionSchema,
+  editConnectionSchema,
+  mergeStructuredIntoConfig,
+  type EditConnectionStructuredPatch,
+} from './edit-connection.schema';
+import type { ConnectionConfigContribution } from '../../../shared/plugins';
 
 describe('mergeStructuredIntoConfig', () => {
   it('writes a new baseUrl into an empty config', () => {
@@ -322,349 +327,79 @@ describe('mergeStructuredIntoConfig — inpostPsModuleType (#767/#1155)', () => 
   });
 });
 
-describe('mergeStructuredIntoConfig — KSeF seller profile (#1223)', () => {
-  it('assembles the nested config.seller shape resolveSeller reads', () => {
+describe('buildEditConnectionSchema / mergeStructuredIntoConfig — plugin contribution seam (#1330)', () => {
+  const contribution: ConnectionConfigContribution = {
+    schemaShape: {
+      acmeToken: z
+        .string()
+        .refine((v) => v === '' || v.startsWith('acme-'), {
+          message: 'Token must start with acme-.',
+        })
+        .optional(),
+    },
+    superRefine: (values, ctx) => {
+      if (values.acmeToken === 'acme-forbidden') {
+        ctx.addIssue({
+          code: z.ZodIssueCode.custom,
+          path: ['acmeToken'],
+          message: 'This token is not allowed.',
+        });
+      }
+    },
+    readConfigToForm: () => ({}),
+    applyToConfig: (config, patch) => {
+      const next = { ...config };
+      if (typeof patch.acmeToken === 'string') {
+        if (patch.acmeToken.length === 0) delete next.token;
+        else next.token = patch.acmeToken;
+      }
+      return next;
+    },
+  };
+
+  it('composes to the plain base schema when no contribution is supplied', () => {
+    const result = buildEditConnectionSchema().safeParse({
+      name: 'Base',
+      configText: '{}',
+    });
+    expect(result.success).toBe(true);
+    // Static base export and no-contribution composition validate identically.
+    expect(editConnectionSchema.safeParse({ name: 'Base', configText: '{}' }).success).toBe(true);
+  });
+
+  it('validates plugin fields through the contributed fragment + superRefine', () => {
+    const schema = buildEditConnectionSchema(contribution);
+    expect(schema.safeParse({ name: 'X', configText: '{}', acmeToken: 'acme-ok' }).success).toBe(
+      true,
+    );
+    expect(schema.safeParse({ name: 'X', configText: '{}', acmeToken: 'nope' }).success).toBe(
+      false,
+    );
+    expect(
+      schema.safeParse({ name: 'X', configText: '{}', acmeToken: 'acme-forbidden' }).success,
+    ).toBe(false);
+  });
+
+  it('leaves the config untouched beyond base clauses when no contribution is supplied', () => {
     const result = mergeStructuredIntoConfig(
-      { env: 'prod' },
-      {
-        sellerNip: '12-3456789-0',
-        sellerName: 'ACME Sp. z o.o.',
-        sellerAddressLine1: 'ul. Przykładowa 1',
-        sellerCity: 'Warszawa',
-        sellerPostalCode: '00-001',
-        sellerCountryIso2: 'pl',
-      },
+      { token: 'keep' },
+      // The synthetic key is not declaration-merged (test-local contribution),
+      // so it needs an explicit widening to the patch type.
+      { acmeToken: 'acme-new' } as EditConnectionStructuredPatch,
     );
-    // NIP normalised to digits, country upper-cased, blank line2 omitted.
-    expect(result.seller).toEqual({
-      nip: '1234567890',
-      name: 'ACME Sp. z o.o.',
-      address: {
-        line1: 'ul. Przykładowa 1',
-        city: 'Warszawa',
-        postalCode: '00-001',
-        countryIso2: 'PL',
-      },
-    });
-    // No flat sellerNip — the canonical location is config.seller.nip.
-    expect(result.sellerNip).toBeUndefined();
-    expect(result.env).toBe('prod');
+    expect(result).toEqual({ token: 'keep' });
   });
 
-  it('preserves untouched seller siblings on a single-field patch', () => {
-    const base = {
-      seller: {
-        nip: '1234567890',
-        name: 'ACME',
-        address: { line1: 'ul. A 1', city: 'Kraków', postalCode: '30-001', countryIso2: 'PL' },
-      },
-    };
-    const result = mergeStructuredIntoConfig(base, { sellerCity: 'Gdańsk' });
-    expect(result.seller).toEqual({
-      nip: '1234567890',
-      name: 'ACME',
-      address: { line1: 'ul. A 1', city: 'Gdańsk', postalCode: '30-001', countryIso2: 'PL' },
-    });
-  });
-
-  it('drops an emptied address object and an emptied seller object', () => {
-    const base = {
-      seller: { address: { city: 'Łódź' } },
-    };
-    const result = mergeStructuredIntoConfig(base, { sellerCity: '' });
-    expect('seller' in result).toBe(false);
-  });
-
-  it('does not touch config.seller when no seller sub-field is on the patch', () => {
-    const base = { seller: { nip: '1234567890' }, env: 'test' };
-    const result = mergeStructuredIntoConfig(base, { ksefEnvironment: 'demo' });
-    expect(result.seller).toEqual({ nip: '1234567890' });
-    expect(result.env).toBe('demo');
-  });
-});
-
-describe('KSeF seller assembly — create/edit parity (#1223)', () => {
-  // Feeds the identical flat seller input through the create path
-  // (buildKsefSellerConfig) and the edit path (mergeStructuredIntoConfig onto an
-  // empty config) and asserts both produce the same nested config.seller. Guards
-  // against the two flows' normalization/assembly drifting apart now that they
-  // share one source (ksef-seller-config).
-  const cases: Array<{ name: string; input: KsefSellerProfileInput }> = [
-    {
-      name: 'full profile with separators + lower-case country',
-      input: {
-        sellerNip: '12-3456789-0',
-        sellerName: '  ACME Sp. z o.o.  ',
-        sellerAddressLine1: ' ul. Przykładowa 1 ',
-        sellerAddressLine2: '',
-        sellerCity: 'Warszawa',
-        sellerPostalCode: '00-001',
-        sellerCountryIso2: 'pl',
-      },
-    },
-    {
-      name: 'NIP only, no address',
-      input: { sellerNip: '1234567890' },
-    },
-    {
-      name: 'name + partial address',
-      input: { sellerName: 'Solo', sellerCity: 'Kraków' },
-    },
-    {
-      name: 'lone country default (hollow profile)',
-      input: { sellerCountryIso2: 'PL' },
-    },
-    {
-      name: 'completely empty',
-      input: {},
-    },
-  ];
-
-  for (const { name, input } of cases) {
-    it(`produces identical config.seller via create and edit paths — ${name}`, () => {
-      const createSeller = buildKsefSellerConfig(input);
-      const editSeller = mergeStructuredIntoConfig({}, input).seller;
-      expect(editSeller).toEqual(createSeller);
-    });
-  }
-
-  it('does not write a hollow seller from a lone PL country default', () => {
-    expect(buildKsefSellerConfig({ sellerCountryIso2: 'PL' })).toBeUndefined();
-    expect('seller' in mergeStructuredIntoConfig({}, { sellerCountryIso2: 'PL' })).toBe(false);
-  });
-});
-
-describe('editConnectionSchema — KSeF postal code validation (#1223)', () => {
-  const base = {
-    name: 'KSeF main',
-    configText: '{"env":"prod"}',
-  };
-
-  it('rejects a malformed PL postal code', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      sellerPostalCode: '1234',
-      sellerCountryIso2: 'PL',
-    });
-    expect(result.success).toBe(false);
-  });
-
-  it('accepts a well-formed PL postal code', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      sellerPostalCode: '00-001',
-      sellerCountryIso2: 'PL',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('allows an empty postal code for incremental save', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      sellerPostalCode: '',
-      sellerCountryIso2: 'PL',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('skips the PL format check for a non-PL country', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      sellerPostalCode: 'SW1A 1AA',
-      sellerCountryIso2: 'GB',
-    });
-    expect(result.success).toBe(true);
-  });
-});
-
-describe('editConnectionSchema — KSeF payment bank account number length (#1311 tech-review)', () => {
-  const base = {
-    name: 'KSeF main',
-    configText: '{"env":"prod"}',
-  };
-
-  it('rejects a bank account number shorter than 10 characters', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentBankAccountNrRb: '123',
-    });
-    expect(result.success).toBe(false);
-  });
-
-  it('accepts a bank account number at the 10-character lower bound', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentBankAccountNrRb: '1234567890',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('allows an empty bank account number for incremental save', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentBankAccountNrRb: '',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('counts the whitespace-stripped length for a conventionally-spaced NRB paste', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      // 26 digits + 6 inner spaces = 32 raw chars; stripped length is what
-      // the persisted wire value carries, so this must pass the 10-34 bound.
-      paymentBankAccountNrRb: '61 1090 1014 0000 0000 9999 9999',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('rejects a spaced paste whose stripped length is under 10', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentBankAccountNrRb: '12 34 56 78 9',
-    });
-    expect(result.success).toBe(false);
-  });
-});
-
-describe('editConnectionSchema — KSeF skonto both-or-neither (#1311 tech-review)', () => {
-  const base = {
-    name: 'KSeF main',
-    configText: '{"env":"prod"}',
-  };
-
-  it('accepts both skonto fields set', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentSkontoConditions: 'paid within 7 days',
-      paymentSkontoAmount: '2%',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('accepts both skonto fields empty', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentSkontoConditions: '',
-      paymentSkontoAmount: '',
-    });
-    expect(result.success).toBe(true);
-  });
-
-  it('anchors the error on the missing amount when only conditions are set', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentSkontoConditions: 'paid within 7 days',
-      paymentSkontoAmount: '',
-    });
-    expect(result.success).toBe(false);
-    expect(result.success ? [] : result.error.issues.map((i) => i.path.join('.'))).toContain(
-      'paymentSkontoAmount',
-    );
-  });
-
-  it('anchors the error on the missing conditions when only the amount is set', () => {
-    const result = editConnectionSchema.safeParse({
-      ...base,
-      paymentSkontoConditions: '',
-      paymentSkontoAmount: '2%',
-    });
-    expect(result.success).toBe(false);
-    expect(result.success ? [] : result.error.issues.map((i) => i.path.join('.'))).toContain(
-      'paymentSkontoConditions',
-    );
-  });
-});
-
-describe('editConnectionSchema — KSeF paymentTermDays sanity cap (#1311 tech-review)', () => {
-  const base = {
-    name: 'KSeF main',
-    configText: '{"env":"prod"}',
-  };
-
-  it('accepts a term at the 999 upper bound', () => {
-    const result = editConnectionSchema.safeParse({ ...base, paymentTermDays: '999' });
-    expect(result.success).toBe(true);
-  });
-
-  it('rejects a term above the 999 sanity cap', () => {
-    const result = editConnectionSchema.safeParse({ ...base, paymentTermDays: '1400' });
-    expect(result.success).toBe(false);
-  });
-});
-
-describe('mergeStructuredIntoConfig — KSeF payment (#1311)', () => {
-  it('assembles the nested config.payment shape resolvePayment reads', () => {
+  it("runs the contribution's applyToConfig as the final assembly pass", () => {
     const result = mergeStructuredIntoConfig(
-      { env: 'prod' },
-      {
-        paymentFormaPlatnosci: '6',
-        paymentBankAccountNrRb: '61 1090 1014 0000 0000 9999 9999',
-        paymentBankAccountBankName: 'Santander',
-        paymentBankAccountSwift: 'WBKPPLPP',
-        paymentTermDays: '14',
-        paymentSkontoConditions: '2% if paid within 7 days',
-        paymentSkontoAmount: '2%',
-      },
+      { baseUrl: 'https://old.example.com', custom: true },
+      { baseUrl: 'https://new.example.com', acmeToken: 'acme-new' } as EditConnectionStructuredPatch,
+      contribution,
     );
-    expect(result.payment).toEqual({
-      formaPlatnosci: '6',
-      bankAccount: {
-        // Whitespace-stripped by `normalizeNrRb` at assembly time — inner
-        // spaces never reach the persisted config or the FA(3) wire.
-        nrRb: '61109010140000000099999999',
-        bankName: 'Santander',
-        swift: 'WBKPPLPP',
-      },
-      paymentTermDays: 14,
-      skonto: { conditions: '2% if paid within 7 days', amount: '2%' },
+    expect(result).toEqual({
+      baseUrl: 'https://new.example.com',
+      custom: true,
+      token: 'acme-new',
     });
-    expect(result.env).toBe('prod');
-  });
-
-  it('assembles formaPlatnosci-only (Gotówka, no bank account)', () => {
-    const result = mergeStructuredIntoConfig({}, { paymentFormaPlatnosci: '1' });
-    expect(result.payment).toEqual({ formaPlatnosci: '1' });
-  });
-
-  it('assembles bankAccount-only when nrRb is set without a payment method', () => {
-    const result = mergeStructuredIntoConfig({}, { paymentBankAccountNrRb: '61109010140000000099999999' });
-    expect(result.payment).toEqual({ bankAccount: { nrRb: '61109010140000000099999999' } });
-  });
-
-  it('preserves untouched payment siblings on a single-field patch', () => {
-    const base = { payment: { formaPlatnosci: '6', paymentTermDays: 14 } };
-    const result = mergeStructuredIntoConfig(base, { paymentSkontoConditions: 'text', paymentSkontoAmount: '5%' });
-    expect(result.payment).toEqual({
-      formaPlatnosci: '6',
-      paymentTermDays: 14,
-      skonto: { conditions: 'text', amount: '5%' },
-    });
-  });
-
-  it('drops an emptied bankAccount object and an emptied payment object', () => {
-    const base = { payment: { bankAccount: { nrRb: '61109010140000000099999999' } } };
-    const result = mergeStructuredIntoConfig(base, { paymentBankAccountNrRb: '' });
-    expect('payment' in result).toBe(false);
-  });
-
-  it('persists an incomplete skonto (missing amount) so per-keystroke sync never drops the first-typed field', () => {
-    // Completeness (conditions+amount both present) is a save-time shape-validator
-    // / issuance-time resolvePayment concern, not a persistence gate (#1311
-    // smoke-test finding).
-    const result = mergeStructuredIntoConfig({}, { paymentSkontoConditions: 'text only' });
-    expect(result.payment).toEqual({ skonto: { conditions: 'text only' } });
-  });
-
-  it('treats a non-numeric paymentTermDays as clearing the field', () => {
-    const base = { payment: { formaPlatnosci: '6', paymentTermDays: 14 } };
-    const result = mergeStructuredIntoConfig(base, { paymentTermDays: 'not-a-number' });
-    expect(result.payment).toEqual({ formaPlatnosci: '6' });
-  });
-
-  it('does not touch config.payment when no payment sub-field is on the patch', () => {
-    const base = { payment: { formaPlatnosci: '6' }, env: 'test' };
-    const result = mergeStructuredIntoConfig(base, { ksefEnvironment: 'demo' });
-    expect(result.payment).toEqual({ formaPlatnosci: '6' });
-    expect(result.env).toBe('demo');
   });
 });
