+++ apps/web/src/plugins/ksef/lib/ksef-nip.ts
@@ -3,11 +3,11 @@
  *
  * Single source of truth for stripping the dashes/spaces an operator may paste
  * into a Polish NIP field. Both the create-path (`ksef-setup.schema.ts`) and the
- * edit-path (`edit-connection.schema.ts` zod transform + `mergeStructuredIntoConfig`)
- * call this so the persisted `config.sellerNip` shape (digits only) can't drift
+ * edit-path (`ksef-connection-config.ts` zod transform + assembly)
+ * call this so the persisted `config.seller.nip` shape (digits only) can't drift
  * between the two flows.
  *
- * @module features/connections/components
+ * @module plugins/ksef/lib
  */
 export function normalizeNip(value: string): string {
   return value.replace(/[\s-]/g, '');
