+++ docs/frontend-architecture.md
@@ -498,6 +498,7 @@ Every slot is optional. A plugin contributes only the affordances its platform a
 | `requiresExternalAuthRedirect` | `boolean` | `CreateConnectionForm` | When true, the inline create form swaps in an Alert linking to the guided wizard (today: Allegro OAuth). Named broadly so non-OAuth redirect flows can opt in. |
 | `getCallbackUrlDefault` | `() => string \| undefined` | `EditConnectionForm` | Default for the OL callback URL field when the connection has none stored. PrestaShop uses `window.location.origin`. |
 | `StructuredConfigSection` | `ComponentType<StructuredConfigSectionProps>` | `EditConnectionForm` | Platform-specific structured-config inputs (PS: shop URL / storefront / shop ID / OL callback / fallback carrier). When absent, the form falls back to raw JSON. |
+| `connectionConfig` | `ConnectionConfigContribution` | `EditConnectionForm`, `mergeStructuredIntoConfig` | The non-render half of structured-config editing (#1330): Zod schema fragment merged into the edit-connection resolver, read-side hydration (`readConfigToForm`), and write-side partial-patch assembly (`applyToConfig`). Field names are declaration-merged into `PluginEditConnectionFields` (mirrors `PluginApiNamespaces`). A platform whose `StructuredConfigSection` registers its own form fields must also contribute this bag. KSeF is the first migrated platform; other platforms' inline fields in `edit-connection.schema.ts` migrate as follow-ups. |
 | `ExtraConfigSection` | `ComponentType<ExtraConfigSectionProps>` | `EditConnectionForm` | Extra section below the structured/raw block (Allegro: GPSR seller defaults). |
 | `CredentialsPanel` | `ComponentType<{ connection }>` | `EditConnectionForm` | Full credentials panel including the rotate-key UI shape that fits the platform's credential model. When absent, the form renders a read-only "Stored securely (managed by integration)" / "Environment variable" affordance. |
 | `ConnectionActions` | `ComponentType<{ connection }>` | `ConnectionActionsPanel` | Extra platform-specific actions on the connection-detail page (PS: "Configure webhooks"). |
